+++ src/main/java/com/supermartijn642/trashcans/filter/IFilterManager.java
@@ -1,7 +1,7 @@
 package com.supermartijn642.trashcans.filter;
 
 import net.minecraft.core.HolderLookup;
-import net.minecraft.nbt.CompoundTag;
+import net.minecraft.nbt.Tag;
 import net.minecraft.world.item.ItemStack;
 
 /**
@@ -11,6 +11,6 @@ public interface IFilterManager {
 
     ItemFilter createFilter(ItemStack stack);
 
-    ItemFilter readFilter(CompoundTag tag, HolderLookup.Provider provider);
+    ItemFilter readFilter(Tag tag, HolderLookup.Provider provider);
 
 }