+++ changelog.md
@@ -1,2 +1,2 @@
 ### Trash Cans 1.0.18
-- Initial release of Trash Cans for Forge 1.21.4
+- Initial release of Trash Cans for Forge 1.21.5